+++ kie-wb-common-stunner/kie-wb-common-stunner-sets/kie-wb-common-stunner-bpmn/kie-wb-common-stunner-bpmn-api/src/test/java/org/kie/workbench/common/stunner/bpmn/definition/property/task/BPMNDiagramTest.java
@@ -97,7 +97,7 @@ public class BPMNDiagramTest {
         BPMNDiagramImpl BPMNDiagramImpl = createValidBpmnDiagram();
         BPMNDiagramImpl.getDiagramSet().setId(new Id(ID_INVALID));
         Set<ConstraintViolation<BPMNDiagramImpl>> violations = this.validator.validate(BPMNDiagramImpl);
-        assertEquals(2,
+        assertEquals(1,
                      violations.size());
     }
