+++ kie-wb-common-stunner/kie-wb-common-stunner-sets/kie-wb-common-stunner-bpmn/kie-wb-common-stunner-bpmn-marshalling/src/test/java/org/kie/workbench/common/stunner/bpmn/client/marshall/service/BPMNClientDiagramServiceTest.java
@@ -367,4 +367,81 @@ public class BPMNClientDiagramServiceTest {
         assertEquals(result, diagram);
         assertEquals(SHAPE_SET_ID, diagram.getMetadata().getShapeSetId());
     }
+
+    @Test
+    public void testGetDiagramTitleWhenIsEmpty() {
+        final String actual = tested.createDiagramTitleFromFilePath("");
+
+        assertEquals("default", actual);
+    }
+
+    @Test
+    public void testGetDiagramTitleWhenIsFileName() {
+        final String fileName = "file.dmn";
+        final String expected = "file";
+
+        final String actual = tested.createDiagramTitleFromFilePath(fileName);
+
+        assertEquals(expected, actual);
+    }
+
+    @Test
+    public void testGetDiagramTitleWhenIsWindowsPath() {
+        final String fileName = "C:\\my path\\folder\\file.dmn";
+        final String expected = "file";
+
+        final String actual = tested.createDiagramTitleFromFilePath(fileName);
+
+        assertEquals(expected, actual);
+    }
+
+    @Test
+    public void testGetDiagramTitleWhenIsUnixPath() {
+        final String fileName = "/users/user/file.dmn";
+        final String expected = "file";
+
+        final String actual = tested.createDiagramTitleFromFilePath(fileName);
+
+        assertEquals(expected, actual);
+    }
+
+    @Test
+    public void testGetDiagramTitleWhenIsWindowsPathMoreWords() {
+        final String fileName = "C:\\my path\\folder\\file a.dmn";
+        final String expected = "file a";
+
+        final String actual = tested.createDiagramTitleFromFilePath(fileName);
+
+        assertEquals(expected, actual);
+    }
+
+    @Test
+    public void testGetDiagramTitleWhenIsUnixPathMoreWords() {
+        final String fileName = "/users/user/file a.dmn";
+        final String expected = "file a";
+
+        final String actual = tested.createDiagramTitleFromFilePath(fileName);
+
+        assertEquals(expected, actual);
+    }
+
+    @Test
+    public void testGetDiagramTitleWhenMoreDotsContained() {
+        final String fileName = "/users/user/file.template.dmn";
+        final String expected = "file.template";
+
+        final String actual = tested.createDiagramTitleFromFilePath(fileName);
+
+        assertEquals(expected, actual);
+    }
+
+    @Test
+    public void testGetDiagramTitleWhenIsNotFileNameOrEmpty() {
+        final String fileName = "Something";
+        final String expected = "Something";
+
+        final String actual = tested.createDiagramTitleFromFilePath(fileName);
+
+        assertEquals(expected, actual);
+    }
 }
