+++ kie-wb-common-stunner/kie-wb-common-stunner-sets/kie-wb-common-stunner-case-mgmt/kie-wb-common-stunner-case-mgmt-api/src/test/java/org/kie/workbench/common/stunner/cm/definition/CaseManagementDiagramTest.java
@@ -86,7 +86,7 @@ public class CaseManagementDiagramTest {
     public void testIDInvalid() {
         tested.getDiagramSet().setId(new Id(ID_INVALID));
         Set<ConstraintViolation<CaseManagementDiagram>> violations = this.validator.validate(tested);
-        assertEquals(2, violations.size());
+        assertEquals(1, violations.size());
     }
 
     @Test
